Rename the Jira scripts to say what they are
file_rtl_sim_jira.py described its trigger rather than its job, and sat next to
create_jira_issue.py, which is the generic client it calls:

  create_jira_issue.py  -> jira_client.py   (Jira REST + de-dup helper)
  file_rtl_sim_jira.py  -> create_jira.py   (files a ticket per failed test)
  test_file_rtl_sim_jira.py -> test_create_jira.py

Updates both workflows, the imports in release_test_report.py and the tests, and
the integration-test notes.
